conexion db a api y get apartamentos (200)

+++ GestionResidenciaApi/Models/Apartamentos.cs
@@ -9,6 +9,6 @@ namespace GestionResidenciaApi.Models
         public int TorreId { get; set; }
         public string Numero { get; set; }
         public string Tipo { get; set; }
-        public int Area { get; set; }
+        public decimal Area { get; set; }
     }
 }
